+++ changes.md
@@ -1,3 +1,5 @@
 #### 1.42.2
 * Added conditional constant settings, allowing WPP_Path, WPP_URL, WPP_Templates, etc. to be overwritten.
+* Added "data-property-pagination-hash" and "data-property-list-hash" attributes to templates.
+* Added "wpp:draw_pagination:settings" filter to allow setting override of pagination element.
 * Added changes.md file.